ai-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
pdfplumber>=0.10.3
nltk>=3.8.1
python-dotenv>=1.0.0
httpx>=0.25.2
google-genai